[BE-3238] Table format support for BodoSQL projections (#4080)
* logical_table_to_table() for table input

* unittest for logical_table_to_table

* handle tuple of array table input case

* implement get_dataframe_all_data()

* inline/optimize get_dataframe_all_data()

* Add alias/DCE/array analysis support for new functions

* distributed analysis for logical_table_to_table

* distributed analysis for get_dataframe_all_data

* optimize out logical_table_to_table

* tests for logical_table_to_table optimization

* move input table to a separate arg

* dead column elimination for logical_table_to_table

* table column dels test

* bug fixes

* remove dead input arguments

* test dead input elimination for logical_table_to_table

* update comment

* handle missing input dist corner cases

+++ bodo/hiframes/dataframe_impl.py
@@ -277,9 +277,9 @@ def overload_dataframe_shape(df):
             return (len(t), num_cols)
 
         return impl
+
     ncols = len(df.columns)
-    # using types.int64 due to lowering error (a Numba tuple handling bug)
-    return lambda df: (len(df), types.int64(ncols))  # pragma: no cover
+    return lambda df: (len(df), ncols)  # pragma: no cover
 
 
 @overload_attribute(DataFrameType, "dtypes")
